moved SampleIModels to its own file
