+++ src/@itwinjs-sandbox/SampleIModels.ts
@@ -1,0 +1,10 @@
+export enum SampleIModels {
+  CoffsHarborDemo = "CoffsHarborDemo",
+  MetroStation = "Metrostation Sample",
+  RetailBuilding = "Retail Building Sample",
+  BayTown = "Bay Town Process Plant",
+  House = "House Sample",
+  Stadium = "Stadium",
+  ExtonCampus = "Exton Campus",
+  Villa = "Villa",
+}
