+++ src/@itwinjs-sandbox/components/imodel-selector/IModelSelector.tsx
@@ -5,7 +5,7 @@
 import React, { FunctionComponent } from "react";
 import { Select } from "@bentley/ui-core";
 import "@bentley/icons-generic-webfont/dist/bentley-icons-generic-webfont.css";
-import { SampleIModels } from "../sample-widget-control/SampleWidgetUiProvider";
+import { SampleIModels } from "@itwinjs-sandbox/SampleIModels";
 
 // The Props and State for this sample component
 interface IModelSelectorProps {
