+++ src/@itwinjs-sandbox/constants.ts
@@ -2,7 +2,8 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-import { SampleIModels } from "@itwinjs-sandbox/components/sample-widget-control/SampleWidgetUiProvider";
+
+import { SampleIModels } from "./SampleIModels";
 
 export const defaultIModelList = [SampleIModels.MetroStation, SampleIModels.RetailBuilding, SampleIModels.BayTown, SampleIModels.House, SampleIModels.Stadium, SampleIModels.ExtonCampus];
 export const defaultIModel = SampleIModels.MetroStation;
