+++ src/@itwinjs-sandbox/index.ts
@@ -6,5 +6,6 @@ export { AuthorizationClient } from "./authentication/AuthorizationClient";
 export { IModelSetup } from "./imodel/IModelSetup";
 export { default3DUiConfig, default2DUiConfig } from "./view/DefaultViewerProps";
 export { ViewSetup } from "./view/ViewSetup";
-export { SampleWidgetUiProvider, SampleIModels } from "./components/sample-widget-control/SampleWidgetUiProvider";
+export { SampleWidgetUiProvider } from "./components/sample-widget-control/SampleWidgetUiProvider";
+export * from "./SampleIModels";
 export * from "./constants";
